Array Assignment
Console app that asks user to select an index of an array of strings and prints string at index to screen.
Asks user to select an index of an array of integers and prints integer at index to screen.
Asks uerser to select index of a list of strings, prints string at index to screen.
Try/Catch block set to to diplsya message for errorenous index selection.